refactor: [#79] implements get-orders use case in nest
