+++ src/modules/checkout/query/database-orders.query.spec.ts
@@ -1,13 +1,13 @@
-import { db } from "../../../Infra/Config";
-import { MysqlConnectionAdapter } from "../../../Infra/Database/MysqlConnectionAdapter";
-import { Email } from "../../@shared/Domain/ValueObject/Email";
-import { Id } from "../../@shared/Domain/ValueObject/Id";
-import { Order } from "../Domain/Entity/Order";
-import { OrderRepositoryDatabase } from "../Repository/Database/OrderRepositoryDatabase";
-import { DatabaseOrdersQuery } from "./DatabaseOrdersQuery";
+import { MysqlConnectionAdapter } from "src/modules/database/connection/mysql/mysql-connection.adapter";
+import { Email } from "src/modules/shared/domain/value-object/email";
+import { Id } from "src/modules/shared/domain/value-object/id";
+import { Order } from "../domain/entity/order";
+import { OrderRepositoryDatabase } from "../repository/database/order.repository";
+import { DatabaseOrdersQuery } from "./database-orders.query";
+import { dbConfig } from "src/modules/database/connection/mysql/config";
 
 describe("DatabaseOrdersQuery tests", () => {
-  const connection = new MysqlConnectionAdapter(db.getConnectionString());
+  const connection = new MysqlConnectionAdapter(dbConfig);
   const orderRepository = new OrderRepositoryDatabase(connection);
   const ordersGateway = new DatabaseOrdersQuery(connection);
