+++ src/modules/checkout/repository/database/coupon.repository.spec.ts
@@ -1,8 +1,8 @@
 import { Coupon } from "../../domain/entity/coupon";
 import { Id } from "src/modules/shared/domain/value-object/id";
-import { MysqlConnectionAdapter } from "../../../database/adapter/mysql/mysql-connection.adapter";
+import { MysqlConnectionAdapter } from "../../../database/connection/mysql/mysql-connection.adapter";
 import { CouponRepositoryDatabase } from "./coupon.repository";
-import { dbConfig } from "src/modules/database/adapter/mysql/mysql-connection.adapter.spec";
+import { dbConfig } from "src/modules/database/connection/mysql/config";
 
 describe("CouponRepository tests", () => {
   const connection = new MysqlConnectionAdapter(dbConfig);
