+++ src/modules/checkout/repository/database/order.repository.spec.ts
@@ -4,9 +4,9 @@ import { Product } from "../../domain/entity/product";
 import { Id } from "src/modules/shared/domain/value-object/id";
 import { Email } from "src/modules/shared/domain/value-object/email";
 import { Dimensions } from "../../domain/entity/dimensions";
-import { MysqlConnectionAdapter } from "../../../database/adapter/mysql/mysql-connection.adapter";
+import { MysqlConnectionAdapter } from "../../../database/connection/mysql/mysql-connection.adapter";
 import { Coupon } from "../../domain/entity/coupon";
-import { dbConfig } from "src/modules/database/adapter/mysql/mysql-connection.adapter.spec";
+import { dbConfig } from "src/modules/database/connection/mysql/config";
 
 describe("Order repository", () => {
   const connection = new MysqlConnectionAdapter(dbConfig);
