+++ src/modules/database/connection/mysql/config.ts
@@ -1,0 +1,10 @@
+import "dotenv/config";
+import { MysqlConnectionOptions } from "./mysql-connection.adapter";
+
+export const dbConfig: MysqlConnectionOptions = {
+  dbHost: process.env.DB_HOST,
+  dbName: process.env.DB_NAME,
+  dbPass: process.env.DB_PASS,
+  dbPort: process.env.DB_PORT,
+  dbUser: process.env.DB_USER,
+};
