+++ src/modules/database/connection/mysql/mysql-connection.adapter.spec.ts
@@ -1,13 +1,5 @@
-import "dotenv/config";
+import { dbConfig } from "./config";
 import { MysqlConnectionAdapter, MysqlConnectionOptions } from "./mysql-connection.adapter";
-
-export const dbConfig: MysqlConnectionOptions = {
-  dbHost: process.env.DB_HOST,
-  dbName: process.env.DB_NAME,
-  dbPass: process.env.DB_PASS,
-  dbPort: process.env.DB_PORT,
-  dbUser: process.env.DB_USER,
-};
 
 describe("Database connection tests", () => {
   it("should connect to a database", async () => {
