+++ src/modules/stock/repository/database/stock-entry.repository.spec.ts
@@ -1,8 +1,8 @@
 import { StockEntry } from "../../domain/entity/stock-entry.entity";
 import { Id } from "src/modules/shared/domain/value-object/id";
-import { MysqlConnectionAdapter } from "../../../database/adapter/mysql/mysql-connection.adapter";
+import { MysqlConnectionAdapter } from "../../../database/connection/mysql/mysql-connection.adapter";
 import { StockEntryRepositoryDatabase } from "./stock-entry.repository";
-import { dbConfig } from "src/modules/database/adapter/mysql/mysql-connection.adapter.spec";
+import { dbConfig } from "src/modules/database/connection/mysql/config";
 
 describe("StockEntryRepositoryDatabase tests", () => {
   const connection = new MysqlConnectionAdapter(dbConfig);
